+++ api/gateway/src/main/java/com/mengyunzhi/waterpurifierfilter/service/IdentityFilterService.java
@@ -1,0 +1,11 @@
+package com.mengyunzhi.waterpurifierfilter.service;
+
+/**
+ * Created by chuhang on 2017/6/19.
+ */
+public interface IdentityFilterService {
+    //sha1加密算法
+    String sha1(String data);
+    //验证请求的信息真伪
+    Boolean isTrue(Long id, String timestamp, String randomString, String encryptionInfo);
+}
