+++ api/gateway/src/main/resources/application.properties
@@ -1,0 +1,5 @@
+zuul.routes.waterPurifier.url=http://localhost:
+
+ribbon.eureka.enabled=false
+
+server.port=8090
