--- api/src/docs/asciidoc/index.adoc
@@ -1,6 +1,0 @@
-include::{generated}/overview.adoc[]
-include::manual_content1.adoc[]
-include::manual_content2.adoc[]
-include::{generated}/paths.adoc[]
-include::{generated}/security.adoc[]
-include::{generated}/definitions.adoc[]
