--- api/src/docs/asciidoc/manual_content2.adoc
@@ -1,3 +1,0 @@
-== Chapter of manual content 2
-
-This is some dummy text
