--- api/src/main/java/com/mengyunzhi/waterPurifierApi/repository/User.java
@@ -1,77 +1,0 @@
-package com.mengyunzhi.waterPurifierApi.repository;
-
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-
-import javax.persistence.*;
-import java.io.Serializable;
-
-/**
- * Created by panjie on 17/5/12.
- * 用户实体
- */
-@Entity
-@ApiModel("用户实体")
-public class User implements Serializable {
-    private static final long serialVersionUID = 1L;
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
-    @ApiModelProperty("姓名") private String name;
-    @ApiModelProperty("用户名") private String username;
-    @ApiModelProperty("密码") private String password;
-
-
-    public User() {
-    }
-
-    public User(String name, String username, String password) {
-        this.name = name;
-        this.username = username;
-        this.password = password;
-    }
-
-    @Override
-    public String toString() {
-        return "User{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", username='" + username + '\'' +
-                ", password='" + password + '\'' +
-                '}';
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
