+++ api/waterpurifier/src/docs/asciidoc/manual_content1.adoc
@@ -1,0 +1,7 @@
+== Chapter of manual content 1
+
+This is some dummy text
+
+=== Sub chapter
+
+Dummy text of sub chapter
