+++ api/waterpurifier/src/main/java/com/mengyunzhi/waterPurifierApi/WaterpurifierapiApplication.java
@@ -1,0 +1,12 @@
+package com.mengyunzhi.waterPurifierApi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WaterpurifierapiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(WaterpurifierapiApplication.class, args);
+	}
+}
